Add files via upload
Data augmentation is a technique used to artificially increase the size of a dataset by applying various transformations to the existing data samples. This technique is commonly used in computer vision tasks, such as image classification, to improve the model's performance and generalization.

Pascal VOC (Visual Object Classes) is a widely used dataset format for object detection and image segmentation tasks. It consists of annotated images along with bounding box coordinates and class labels.

To perform data augmentation on images in the Pascal VOC format, you can use libraries such as OpenCV, TensorFlow, or Keras. Here's an example using the popular library, OpenCV,